Tools package and scene cache
Some of the existing helper functions and classes (e.g. app timer, draw rounded rectangle, etc.) have been moved to a new folder called `tools`. These include some of the functions that were originally in `shared.py`, but they can still be imported from the `shared.py` module for now.

A new "scene cache" system in `app_main.py` that enables saving previous scenes into a cache when changing into another scene, so that the program does not have to load the scenes over and over again. This is done in hopes of reducing lag by a bit.

The bug where a card is blank after being revealed may have been fixed.
