+++ app/__init__.py
@@ -1,3 +1,3 @@
 from . import app_main
 from . import shared
-from . import app_timer
+from .tools import app_timer
