+++ app/tools/colors.py
@@ -1,0 +1,33 @@
+import colorsys
+import random
+
+
+def hsv_factor(rgb: tuple or str, hf=0, sf=1, vf=1) -> tuple:
+    """
+    Takes a 24 bit RGB value and changes it according to the given HSV factors (hue, saturation, and value)
+
+    :param rgb: The RGB value can be in tuple (e.g. (255, 255, 255)) or a string with the "#RRGGBB" format
+
+    :param hf: Hue factor
+    :param sf: Saturation factor
+    :param vf: Value (brightness) factor
+    """
+    if type(rgb) is tuple and len(rgb) == 4:
+        return hsv_factor(rgb[:-1], hf, sf, vf) + (rgb[-1],)
+
+    r, g, b = (x / 255.0 for x in rgb)
+    h, s, v = colorsys.rgb_to_hsv(r, g, b)
+
+    new_hsv = h + hf, s * sf, v * vf
+    new_hsv = tuple(max(0, min(1, x)) for x in new_hsv)  # Clamp values between 0-1
+    new_rgb = colorsys.hsv_to_rgb(*new_hsv)
+    int_rgb = tuple(int(x * 255) for x in new_rgb)  # Convert RGB values from 0.0 - 1.0 to 0 - 255
+
+    return int_rgb
+
+
+def rand_color() -> tuple:
+    """
+    Generates a random color in an (R, G, B) tuple format.
+    """
+    return tuple(random.randrange(256) for _ in range(3))
